I installed gtest in /usr/share and updated the Eclipse project to reflect that
The significance of this is that modifying the Eclipse properties to reflect the location
of gtest was the most complex part of getting up and running with the test cases.  All one
needs to do now is install gtest in /usr/share rather than their home directory (or any
place else) and everything should work perfectly.